requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
strands-agents[anthropic]
strands-agents-tools
PyPDF2
python-multipart